+++ src/main/java/com/fastfoodsm/fastfood/model/Food.java
@@ -41,7 +41,7 @@ public class Food implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Size(max = 255)
     @Column(name = "foodName")
     private String foodName;
@@ -63,15 +63,15 @@ public class Food implements Serializable {
     public Food() {
     }
 
-    public Food(Integer id) {
+    public Food(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
